refactored logo on /about and / to use only Logo component in home folder
