+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trend-forge",
-  "version": "1.3.2",
+  "version": "1.3.3",
   "license": "MIT",
   "private": true,
   "scripts": {
